requirements.txt
python-dateutil
requests

uvicorn
fastapi
pydantic